+++ termbox-bindings-hs/src/Termbox/Bindings/Hs/Internal/EventMod.hs
@@ -13,7 +13,12 @@ import qualified Termbox.Bindings.C
 -- | An event modifier.
 newtype Tb_event_mod
   = Tb_event_mod Word8
-  deriving stock (Eq, Ord, Show)
+  deriving stock (Eq, Ord)
+
+instance Show Tb_event_mod where
+  show = \case
+    TB_MOD_ALT -> "TB_MOD_ALT"
+    TB_MOD_MOTION -> "TB_MOD_MOTION"
 
 pattern TB_MOD_ALT :: Tb_event_mod
 pattern TB_MOD_ALT <-
@@ -26,3 +31,6 @@ pattern TB_MOD_MOTION <-
   ((== Tb_event_mod Termbox.Bindings.C._TB_MOD_MOTION) -> True)
   where
     TB_MOD_MOTION = Tb_event_mod Termbox.Bindings.C._TB_MOD_MOTION
+
+-- N.B. This requires Tb_event_mod to remain abstract
+{-# COMPLETE TB_MOD_ALT, TB_MOD_MOTION #-}