+++ termbox-bindings-hs/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## [1.0.0] - Unreleased
 
-- Make `Show` instance of `Tb_key` not call `error` on unknown keys
 - Rework `Tb_attr` and `Tb_color` into `Tb_color_and_attrs`
+- Make `mod` field of `Tb_event` a `Maybe Tb_event_mod` rather than a `Tb_event_mod`
+- Make `Show` instance of `Tb_key` not call `error` on unknown keys
+- Hide representation of `Tb_event_mod`
+- Improve `Show` instance of `Tb_event_mod`
 - Delete `tb_attr`
 
 ## [0.1.1] - November 5, 2023